update : branch, purpose 유효성 검사 위치 수정
