+++ src/main/java/kr/co/conceptbe/notification_setting/domain/vo/NotificationSettingPurposes.java
@@ -5,6 +5,7 @@ import jakarta.persistence.Embeddable;
 import jakarta.persistence.OneToMany;
 import java.util.Set;
 import kr.co.conceptbe.notification_setting.domain.IdeaNotificationSetting;
+import kr.co.conceptbe.notification_setting.exception.InvalidPurposeException;
 import kr.co.conceptbe.purpose.domain.Purpose;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -35,6 +36,8 @@ public class NotificationSettingPurposes {
             IdeaNotificationSetting ideaNotificationSetting,
             Set<Purpose> purposes
     ) {
+        validatePurpose(purposes);
+
         List<NotificationSettingPurpose> notificationSettingPurposes = purposes.stream()
                 .map(purpose -> NotificationSettingPurpose.of(ideaNotificationSetting, purpose))
                 .toList();
@@ -42,12 +45,23 @@ public class NotificationSettingPurposes {
         return new NotificationSettingPurposes(notificationSettingPurposes);
     }
 
-    public void update(IdeaNotificationSetting ideaNotificationSetting, Set<Purpose> purposes) {
+    public void update(
+            IdeaNotificationSetting ideaNotificationSetting,
+            Set<Purpose> purposes
+    ) {
+        validatePurpose(purposes);
+
         notificationSettingPurposes.clear();
         notificationSettingPurposes.addAll(
             purposes.stream()
                 .map(purpose -> NotificationSettingPurpose.of(ideaNotificationSetting, purpose))
                 .toList()
         );
+    }
+
+    private static void validatePurpose(Set<Purpose> purposes) {
+        if (purposes.isEmpty()) {
+            throw new InvalidPurposeException();
+        }
     }
 }
